backend/my_app/urls.py
from django.urls import path
from rest_framework.routers import DefaultRouter
from .views import UserViewSet

router = DefaultRouter()
router.register(r'users', UserViewSet, basename='user')  # Change from products to users

urlpatterns = router.urls
